requirements.txt
beautifulsoup4
emoji==0.5.4
google-api-core==1.16.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-translate==2.0.0
google-pasta==0.2.0
googleapis-common-protos==1.51.0
numpy
scikit-learn
scipy
text-unidecode
torch
tornado
nltk
requests
jieba